refactor: move FilePatternList to atoms
