+++ src/components/atoms/FilePatternList/FilePatternList.styled.tsx
