+++ src/components/atoms/FilePatternList/FilePatternListItem.styled.tsx
