+++ src/components/atoms/FilePatternList/FilePatternListItem.tsx
