+++ src/components/atoms/index.ts
@@ -2,6 +2,7 @@ export {default as CopyButton} from './CopyButton';
 export {default as ContextMenu} from './ContextMenu';
 export {default as Dots} from './Dots';
 export {default as FileExplorer} from './FileExplorer';
+export {default as FilePatternList} from './FilePatternList';
 export {default as HotkeyLabel} from './HotkeyLabel';
 export {default as Icon} from './Icon';
 export {default as IconWithPopover} from './IconWithPopover';
